Improve Protocol design.
The Protocol class no longer requires message types to have associated
MessageReaders. Instead, a Connection class has been created, which
handles sending and recieving the bytes of each message over the
network, including prepending each message with a message size, and
reading the entire message (not including the size header) into an
appropriately sized byte[], which is passed to a Protocol instance.