Implement TestListRoomsRequest, update ClientServerSeq.
This commit adds a clearRecieved method to ClientServerSeq. This method
can be inserted into a method chain to clear the queued recieved
messages (useful for when an unknown number of messages is expected).

This commit also adds support for passing in a null message to .expect
and .expect2 of ClientServerSeq, which enables the expectation of a
particular message type, but without any expectations of the message
content.